Nhutlq

+++ src/main/java/com/main/entity/Discounts.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.Collection;
@@ -25,7 +26,7 @@ public class Discounts {
 
     @Basic
     @Column(name = "discount_cost", nullable = true, precision = 0)
-    private Integer discountCost;
+    private BigDecimal discountCost;
 
     @Basic
     @Column(name = "quantity", nullable = true)
@@ -50,35 +51,4 @@ public class Discounts {
     @OneToMany(mappedBy = "discountsByDiscountId")
     @JsonManagedReference
     private Collection<Orders> ordersById;
-
-//    @Override
-//    public boolean equals(Object o) {
-//        if (this == o) return true;
-//        if (o == null || getClass() != o.getClass()) return false;
-//
-//        Discounts that = (Discounts) o;
-//
-//        if (quantity != null ? !quantity.equals(that.quantity) : that.quantity != null) return false;
-//        if (id != null ? !id.equals(that.id) : that.id != null) return false;
-//        if (startUse != null ? !startUse.equals(that.startUse) : that.startUse != null) return false;
-//        if (endUse != null ? !endUse.equals(that.endUse) : that.endUse != null) return false;
-//        if (discountDescription != null ? !discountDescription.equals(that.discountDescription) : that.discountDescription != null)
-//            return false;
-//        if (isActive != that.isActive) return false;
-//
-//        return true;
-//    }
-//
-//    @Override
-//    public int hashCode() {
-//        int result = id != null ? id.hashCode() : 0;
-//        result = 31 * result + (quantity != null ? quantity.hashCode() : 0);
-//        result = 31 * result + (startUse != null ? startUse.hashCode() : 0);
-//        result = 31 * result + (endUse != null ? endUse.hashCode() : 0);
-//        result = 31 * result + (discountDescription != null ? discountDescription.hashCode() : 0);
-//        result = 31 * result + (isActive ? 1 : 0);
-//        return result;
-//    }
-
-
 }
